# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Subscriptions API
  version: 1.0.0
  description: Сервис для агрегации данных о подписках пользователей.

paths:
  /subscriptions:
    post:
      summary: Создать новую подписку
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        '201':
          description: Подписка успешно создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
    get:
      summary: Получить список всех подписок
      responses:
        '200':
          description: Список подписок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'

  /subscriptions/{id}:
    get:
      summary: Получить подписку по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные подписки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
    put:
      summary: Обновить данные подписки
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        '200':
          description: Успешно обновлено
    delete:
      summary: Удалить подписку
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешно удалено

  /subscriptions/total:
    get:
      summary: Суммарная стоимость за период
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: service_name
          in: query
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: true
          schema:
            type: string
            example: "07-2025"
      responses:
        '200':
          description: Итоговая сумма
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_price:
                    type: integer

components:
  schemas:
    Subscription:
      type: object
      required:
        - service_name
        - price
        - user_id
        - start_date
      properties:
        id:
          type: integer
          readOnly: true
        service_name:
          type: string
          example: "Yandex Plus"
        price:
          type: integer
          example: 400
        user_id:
          type: string
          format: uuid
          example: "60601fee-2bf1-4721-ae6f-7636e79a0cba"
        start_date:
          type: string
          example: "07-2025"
        end_date:
          type: string
          example: "07-2026"
          nullable: true